move LifecycleModel parameters dependence out of HARK library in to the example, fixes #584
